requirements.txt
tensorflow==2.2.0
keras==2.4.3
cv2==4.2.0
Pillow==7.1.2
numpy==1.19.0